+++ lib/KR.ml
@@ -20,20 +20,201 @@ let src = Logs.Src.create "okra.KR"
 
 module Log = (val Logs.src_log src : Logs.LOG)
 
-type id = New_KR | No_KR | ID of string
+let compare_no_case x y =
+  let x = String.uppercase_ascii x in
+  let y = String.uppercase_ascii y in
+  String.compare x y
 
-let pp_id ppf = function
-  | New_KR -> Fmt.pf ppf "New KR"
-  | No_KR -> Fmt.pf ppf "No KR"
-  | ID s -> Fmt.pf ppf "%s" s
+module Meta = struct
+  type t = Community | Hack | Learning | Leave | Management | Meet | Onboard
+
+  let pp ppf = function
+    | Community -> Fmt.pf ppf "Community"
+    | Hack -> Fmt.pf ppf "Hack"
+    | Learning -> Fmt.pf ppf "Learning"
+    | Leave -> Fmt.pf ppf "Leave"
+    | Management -> Fmt.pf ppf "Management"
+    | Meet -> Fmt.pf ppf "Meet"
+    | Onboard -> Fmt.pf ppf "Onboard"
+
+  let of_string s =
+    match String.lowercase_ascii s with
+    | "community" -> Some Community
+    | "hack" -> Some Hack
+    | "learning" -> Some Learning
+    | "leave" -> Some Leave
+    | "management" -> Some Management
+    | "meet" -> Some Meet
+    | "onboard" -> Some Onboard
+    | _ -> None
+
+  let compare x y =
+    String.compare (Fmt.to_to_string pp x) (Fmt.to_to_string pp y)
+end
+
+module Work = struct
+  module Id = struct
+    type t = New_KR | No_KR | ID of string
+
+    let pp ppf = function
+      | New_KR -> Fmt.pf ppf "New KR"
+      | No_KR -> Fmt.pf ppf "No KR"
+      | ID s -> Fmt.pf ppf "%s" s
+
+    let equal a b =
+      match (a, b) with
+      | New_KR, New_KR | No_KR, No_KR -> true
+      | ID id1, ID id2 -> compare_no_case id1 id2 = 0
+      | _ -> false
+
+    let okr_re = Str.regexp "\\([a-zA-Z#]+[0-9]+\\)"
+    (* Legacy KR: (KR12) *)
+    (* GitHub WI: (12) *)
+
+    let of_string s =
+      match String.lowercase_ascii s with
+      | "new kr" | "new okr" | "new wi" -> Some New_KR
+      | "no kr" | "no okr" | "no wi" -> Some No_KR
+      | _ -> (
+          match Str.string_match okr_re s 0 with
+          | false -> None
+          | true ->
+              let id = String.trim (Str.matched_group 1 s) in
+              Some (ID id))
+
+    let merge ~title x y =
+      match (x, y) with
+      | ID x, ID y ->
+          assert (compare_no_case x y = 0);
+          ID x
+      | ID x, _ | _, ID x -> ID x
+      | No_KR, No_KR -> No_KR
+      | New_KR, New_KR -> New_KR
+      | No_KR, New_KR | New_KR, No_KR ->
+          Fmt.failwith
+            "Mismatch between KR kinds. Same title was used with both No KR \
+             and New KR. Title: %s"
+            title
+  end
+
+  type t = { id : Id.t; title : string; quarter : Quarter.t option }
+
+  let dump =
+    let open Fmt.Dump in
+    record
+      [
+        field "title" (fun t -> t.title) string;
+        field "id" (fun t -> t.id) Id.pp;
+        field "quarter" (fun t -> t.quarter) (Fmt.option Quarter.pp);
+      ]
+
+  let pp ppf x = Format.fprintf ppf "%s (%a)" x.title Id.pp x.id
+  let v ~title ~id ~quarter = { title; id; quarter }
+
+  let compare a b =
+    match (a.id, b.id) with
+    | ID a, ID b -> compare_no_case a b
+    | _ -> compare_no_case a.title b.title
+
+  let merge x y =
+    let title =
+      match (x.title, y.title) with
+      | "", s | s, "" -> s
+      | x, y ->
+          if compare_no_case x y <> 0 then
+            Log.warn (fun l -> l "Conflicting titles:\n- %S\n- %S" x y);
+          x
+    in
+    let quarter =
+      match (x.quarter, y.quarter) with
+      | None, q | q, None -> q
+      | Some x, Some y ->
+          if Quarter.compare x y <> 0 then
+            Log.warn (fun l ->
+                l "KR %S appears in two quarters:\n- %a\n- %a" title Quarter.pp
+                  x Quarter.pp y);
+          Some x
+    in
+    let id = Id.merge ~title x.id y.id in
+    { title; id; quarter }
+end
+
+module Heading = struct
+  type t = Meta of Meta.t | Work of string * Work.Id.t option
+
+  let pp ppf = function
+    | Meta x -> Fmt.pf ppf "%a" Meta.pp x
+    | Work (s, Some id) -> Fmt.pf ppf "%s (%a)" s Work.Id.pp id
+    | Work (s, None) -> Fmt.pf ppf "%s" s
+
+  let of_string s =
+    let title, id =
+      match String.split_on_char '(' s with
+      | [] -> (None, None)
+      | title :: rest -> (
+          match rest with
+          | [] -> (Some title, None)
+          | id :: _ -> (
+              match String.split_on_char ')' id with
+              | [] -> (Some title, Some id)
+              | id :: _ -> (Some title, Some id)))
+    in
+    match title with
+    | Some title -> (
+        let title = String.trim title in
+        match Meta.of_string title with
+        | Some m -> Meta m
+        | None -> (
+            match id with
+            | Some id -> (
+                let id = String.trim id in
+                match Work.Id.of_string id with
+                | Some id -> Work (title, Some id)
+                | None -> Work (title, None))
+            | None -> Work (title, None)))
+    | None -> Work (s, None)
+end
+
+module Kind = struct
+  type t = Meta of Meta.t | Work of Work.t
+
+  let pp ppf = function Meta x -> Meta.pp ppf x | Work x -> Work.pp ppf x
+  let dump ppf = function Meta x -> Meta.pp ppf x | Work x -> Work.dump ppf x
+
+  let merge x y =
+    match (x, y) with
+    | Meta x, Meta y ->
+        assert (Meta.compare x y = 0);
+        Meta x
+    | Work x, Work y -> Work (Work.merge x y)
+    | Meta _, Work _ -> assert false
+    | Work _, Meta _ -> assert false
+
+  let compare a b =
+    let meta = Fmt.to_to_string Meta.pp in
+    let work = Fmt.to_to_string Work.pp in
+    match (a, b) with
+    | Meta a, Meta b -> String.compare (meta a) (meta b)
+    | Meta a, Work b -> String.compare (meta a) (work b)
+    | Work a, Meta b -> String.compare (work a) (meta b)
+    | Work a, Work b -> Work.compare a b
+end
+
+module Id = struct
+  type t = Kind.t
+
+  let pp ppf = function
+    | Kind.Meta x -> Meta.pp ppf x
+    | Kind.Work x -> Work.pp ppf x
+
+  let compare = Kind.compare
+end
 
 type t = {
   counter : int;
+  kind : Kind.t;
   project : string;
   objective : string;
-  title : string;
-  id : id;
-  quarter : Quarter.t option;
   time_entries : (string * Time.t) list list;
   time_per_engineer : (string, Time.t) Hashtbl.t;
   work : Item.t list list;
@@ -46,7 +227,7 @@ let counter =
     incr c;
     i
 
-let v ~project ~objective ~title ~id ~quarter ~time_entries work =
+let v ~kind ~project ~objective ~time_entries work =
   let counter = counter () in
   (* Sum time per engineer *)
   let time_per_engineer =
@@ -61,27 +242,16 @@ let v ~project ~objective ~title ~id ~quarter ~time_entries work =
       time_entries;
     tbl
   in
-  {
-    counter;
-    project;
-    objective;
-    title;
-    id;
-    quarter;
-    time_entries;
-    time_per_engineer;
-    work;
-  }
+  { counter; kind; project; objective; time_entries; time_per_engineer; work }
 
 let dump =
   let open Fmt.Dump in
   record
     [
       field "counter" (fun t -> t.counter) Fmt.int;
+      field "kind" (fun t -> t.kind) Kind.dump;
       field "project" (fun t -> t.project) string;
       field "objective" (fun t -> t.objective) string;
-      field "title" (fun t -> t.title) string;
-      field "id" (fun t -> t.id) pp_id;
       field "time_entries"
         (fun t -> t.time_entries)
         (list (list (pair string Time.pp)));
@@ -91,34 +261,16 @@ let dump =
       field "work" (fun t -> t.work) (list (list Item.dump));
     ]
 
-let compare_no_case x y =
-  let x = String.uppercase_ascii x in
-  let y = String.uppercase_ascii y in
-  String.compare x y
-
-let equal_id a b =
-  match (a, b) with
-  | New_KR, New_KR | No_KR, No_KR -> true
-  | ID id1, ID id2 -> compare_no_case id1 id2 = 0
-  | _ -> false
-
 let merge x y =
   let counter = x.counter in
-  let title =
-    match (x.title, y.title) with
-    | "", s | s, "" -> s
-    | x, y ->
-        if compare_no_case x y <> 0 then
-          Log.warn (fun l -> l "Conflicting titles:\n- %S\n- %S" x y);
-        x
-  in
+  let kind = Kind.merge x.kind y.kind in
   let project =
     match (x.project, y.project) with
     | "", s | s, "" -> s
     | x, y ->
         if compare_no_case x y <> 0 then
           Log.warn (fun l ->
-              l "KR %S appears in two projects:\n- %S\n- %S" title x y);
+              l "KR %a appears in two projects:\n- %S\n- %S" Kind.pp kind x y);
         x
   in
   let objective =
@@ -127,32 +279,8 @@ let merge x y =
     | x, y ->
         if compare_no_case x y <> 0 then
           Log.warn (fun l ->
-              l "KR %S appears in two objectives:\n- %S\n- %S" title x y);
+              l "KR %a appears in two objectives:\n- %S\n- %S" Kind.pp kind x y);
         x
-  in
-  let quarter =
-    match (x.quarter, y.quarter) with
-    | None, q | q, None -> q
-    | Some x, Some y ->
-        if Quarter.compare x y <> 0 then
-          Log.warn (fun l ->
-              l "KR %S appears in two quarters:\n- %a\n- %a" title Quarter.pp x
-                Quarter.pp y);
-        Some x
-  in
-  let id =
-    match (x.id, y.id) with
-    | ID x, ID y ->
-        assert (compare_no_case x y = 0);
-        ID x
-    | ID x, _ | _, ID x -> ID x
-    | No_KR, No_KR -> No_KR
-    | New_KR, New_KR -> New_KR
-    | No_KR, New_KR | New_KR, No_KR ->
-        Fmt.failwith
-          "Mismatch between KR kinds. Same title was used with both No KR and \
-           New KR. Title: %s"
-          title
   in
   let time_entries = x.time_entries @ y.time_entries in
   let time_per_engineer =
@@ -168,22 +296,9 @@ let merge x y =
     t
   in
   let work = x.work @ y.work in
-  {
-    counter;
-    project;
-    objective;
-    title;
-    id;
-    quarter;
-    time_entries;
-    time_per_engineer;
-    work;
-  }
+  { counter; kind; project; objective; time_entries; time_per_engineer; work }
 
-let compare a b =
-  match (a.id, b.id) with
-  | ID a, ID b -> compare_no_case a b
-  | _ -> compare_no_case a.title b.title
+let compare a b = Kind.compare a.kind b.kind
 
 let make_engineer ~time (e, d) =
   if time then Fmt.str "%a (%a)" (User.pp ~with_link:false) e Time.pp d
@@ -209,38 +324,42 @@ let make_time_entries t =
   Item.[ Paragraph (Text (String.concat ", " (List.map aux t))) ]
 
 let update_from_master_db t db =
-  let update (orig_kr : t) (db_kr : Masterdb.elt_t option) =
+  let update (orig_kr : t) (orig_work : Work.t) (db_kr : Masterdb.elt_t option)
+      =
     match db_kr with
     | None ->
-        if orig_kr.id = New_KR then
-          Log.warn (fun l -> l "KR ID not found for new KR %S" orig_kr.title);
+        if orig_work.id = New_KR then
+          Log.warn (fun l -> l "KR ID not found for new KR %S" orig_work.title);
         orig_kr
     | Some db_kr ->
-        if orig_kr.id = No_KR then
+        if orig_work.id = No_KR then
           Log.warn (fun l ->
               l "KR ID updated from \"No KR\" to %S:\n- %S\n- %S" db_kr.id
-                orig_kr.title db_kr.title);
-        let kr =
+                orig_work.title db_kr.title);
+        let work =
           {
-            orig_kr with
-            id = ID db_kr.printable_id;
+            Work.id = ID db_kr.printable_id;
             title = db_kr.title;
-            objective = db_kr.objective;
             quarter = db_kr.quarter;
           }
+        in
+        let kr =
+          { orig_kr with kind = Work work; objective = db_kr.objective }
         in
         (* show the warnings *)
         ignore (merge orig_kr kr);
         kr
   in
-
-  match t.id with
-  | ID id ->
-      let db_kr = Masterdb.find_kr_opt db id in
-      update t db_kr
-  | _ ->
-      let db_kr = Masterdb.find_title_opt db t.title in
-      update t db_kr
+  match t.kind with
+  | Meta _ -> t
+  | Work x -> (
+      match x.id with
+      | ID id ->
+          let db_kr = Masterdb.find_kr_opt db id in
+          update t x db_kr
+      | _ ->
+          let db_kr = Masterdb.find_title_opt db x.title in
+          update t x db_kr)
 
 let items ?(show_time = true) ?(show_time_calc = false) ?(show_engineers = true)
     kr =
@@ -262,7 +381,7 @@ let items ?(show_time = true) ?(show_time_calc = false) ?(show_engineers = true)
       ( Bullet '-',
         [
           [
-            Paragraph (Text (Fmt.str "%s (%a)" kr.title pp_id kr.id));
+            Paragraph (Text (Fmt.str "%a" Kind.pp kr.kind));
             List (Bullet '-', items :: kr.work);
           ];
         ] );