+++ lib/aggregate.mli
@@ -20,7 +20,7 @@ type time_per_engineer = (string, Time.t) Hashtbl.t
 val by_kr :
   ?include_krs:string list ->
   Report.t ->
-  (string * KR.id, Time.t * time_per_engineer) Hashtbl.t
+  (KR.Id.t, Time.t * time_per_engineer) Hashtbl.t
 
 val by_objective :
   ?include_krs:string list ->