+++ lib/lint.mli
@@ -26,7 +26,7 @@ type lint_error =
   | No_project_found of int option * string
   | Not_all_includes of string list
   | Invalid_markdown_in_work_items of int option * string
-  | Invalid_quarter of KR.t
+  | Invalid_quarter of KR.Work.t
 
 type lint_result = (unit, lint_error list) result