+++ test/cram/cat/merge.t
@@ -10,6 +10,8 @@ Engineer reports are aggregated correctly
   >   - @eng1 (1.5 days)
   >   - Some work
   > 
+  > - Leave
+  >   - @eng1 (2 days)
   > EOF
 
   $ cat > eng2.md << EOF
@@ -18,6 +20,9 @@ Engineer reports are aggregated correctly
   > - A Kr (KR123)
   >   - @eng2 (1.5 days)
   >   - Some other work
+  > 
+  > - Leave
+  >   - @eng2 (2 days)
   > 
   > EOF
 
@@ -32,6 +37,8 @@ Engineer reports are aggregated correctly
   >   - @eng3 (1.5 days)
   >   - Some work
   > 
+  > - Leave
+  >   - @eng3 (2 days)
   > EOF
 
   $ cat > okrs.csv << EOF
@@ -49,6 +56,9 @@ Behaviour without a database
     - @eng1 (1.5 days), @eng2 (1.5 days)
     - Some work
     - Some other work
+  
+  - Leave
+    - @eng1 (2 days), @eng2 (2 days)
 
   $ okra cat --engineer --append-to=agg.md eng3.md > agg2.md && cat agg2.md
   # Last Week
@@ -62,11 +72,17 @@ Behaviour without a database
   - A Different Kr (KR124)
     - @eng3 (1.5 days)
     - Some work
+  
+  - Leave
+    - @eng1 (2 days), @eng2 (2 days), @eng3 (2 days)
 
 Behaviour with a database
 
   $ cat eng1.md eng2.md | okra cat --okr-db=okrs.csv --engineer > agg.md && cat agg.md
   # Last Week
+  
+  - Leave
+    - @eng1 (2 days), @eng2 (2 days)
   
   ## Actual objective
   
@@ -77,6 +93,9 @@ Behaviour with a database
 
   $ okra cat --okr-db=okrs.csv --engineer --append-to=agg.md eng3.md > agg2.md && cat agg2.md
   # Last Week
+  
+  - Leave
+    - @eng1 (2 days), @eng2 (2 days), @eng3 (2 days)
   
   ## Actual objective