+++ test/cram/lint/engineer.t
@@ -62,11 +62,11 @@ Errors in include section are detected even if the rest is ignored.
   > EOF
   [ERROR(S)]: <stdin>
   
-  In KR "This is a KR (KRID)":
+  In KR "This is a KR":
     No time entry found. Each KR must be followed by '- @... (x days)'
   [ERROR(S)]: <stdin>
   
-  In KR "This is a KR (KRID)":
+  In KR "This is a KR":
     No KR ID found. WIs should be in the format "This is a WI (#123)", where 123 is the WI issue ID. Legacy KRs should be in the format "This is a KR (PLAT123)", where PLAT123 is the KR ID. For WIs that don't have an ID yet, use "New WI" and for work without a WI use "No WI".
   [1]
 
@@ -95,7 +95,7 @@ Only "No KR" and "New KR" are supported for KR's without identifiers
   Invalid total time found for eng1 (reported 1 day, expected 5 days).
   [ERROR(S)]: <stdin>
   
-  In KR "This is a KR (Off KR)":
+  In KR "This is a KR":
     No KR ID found. WIs should be in the format "This is a WI (#123)", where 123 is the WI issue ID. Legacy KRs should be in the format "This is a KR (PLAT123)", where PLAT123 is the KR ID. For WIs that don't have an ID yet, use "New WI" and for work without a WI use "No WI".
   [1]
 
@@ -119,7 +119,7 @@ Only "No KR" and "New KR" are supported for KR's without identifiers
   > EOF
   [ERROR(S)]: <stdin>
   
-  In KR "This is a KR":
+  In KR "This is a KR (KR1)":
     Invalid time entry "@eng1 (1.1 day)" found. Format is '- @eng1 (x days), @eng2 (y days)'
     where x and y must be divisible by 0.5
   [1]
@@ -144,7 +144,7 @@ Only "No KR" and "New KR" are supported for KR's without identifiers
   > EOF
   [ERROR(S)]: <stdin>
   
-  In KR "This is a KR":
+  In KR "This is a KR (KR1)":
     Invalid time entry "@eng1 ( day)" found. Format is '- @eng1 (x days), @eng2 (y days)'
     where x and y must be divisible by 0.5
   [1]