+++ test/cram/lint/features.t
@@ -9,7 +9,7 @@ Lint can read from a file:
   $ okra lint err.md
   [ERROR(S)]: err.md
   
-  In KR "Everything is great":
+  In KR "Everything is great (E1)":
     No time entry found. Each KR must be followed by '- @... (x days)'
   [1]
 
@@ -18,7 +18,7 @@ It can also read from stdin:
   $ okra lint < err.md
   [ERROR(S)]: <stdin>
   
-  In KR "Everything is great":
+  In KR "Everything is great (E1)":
     No time entry found. Each KR must be followed by '- @... (x days)'
   [1]
 
@@ -49,12 +49,12 @@ When errors are found in several files, they are all printed:
   $ okra lint err.md err2.md
   [ERROR(S)]: err2.md
   
-  In KR "Everything is great":
+  In KR "Everything is great (E1)":
     Invalid time entry "@a" found. Format is '- @eng1 (x days), @eng2 (y days)'
     where x and y must be divisible by 0.5
   [ERROR(S)]: err.md
   
-  In KR "Everything is great":
+  In KR "Everything is great (E1)":
     No time entry found. Each KR must be followed by '- @... (x days)'
   [1]